Revert "LPS-73481 Moving upgrade processes due to new features to next minor release upgrade"
This reverts commit a427e23e57f488c4ea035416e1ed0ec0b0c67255.
