+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_4.java
@@ -20,6 +20,7 @@ import com.liferay.portal.upgrade.v7_0_4.UpgradeContact;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeEmailAddress;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeExpando;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeMBMailingList;
+import com.liferay.portal.upgrade.v7_0_4.UpgradeSchema;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeUser;
 
 /**
@@ -34,6 +35,8 @@ public class UpgradeProcess_7_0_4 extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
+		upgrade(UpgradeSchema.class);
+
 		upgrade(UpgradeContact.class);
 		upgrade(UpgradeEmailAddress.class);
 		upgrade(UpgradeExpando.class);
